Hyperlinks + integration with Fungus
I am just receiving the index of the hyperlink and storing it inside the fungus Flowchart, seems to be the most elegant way.
What is not as elegant is the styling of the hyperlinks lol, for now the only way we cn do it is through rich text tags, which looks kinda clunky (check the flowchat on the Test scene to see what i mean lol). Maybe I can find a way to make custom tags, then it would look a bit more elegant